# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Verisafe service API",
        "title": "Verisafe API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Open Crafts Interactive",
            "url": "https://opencrafts.io/about",
            "email": "developers@opencrafts.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "host": "qaverisafe.opencrafts.io",
    "basePath": "/",
    "paths": {
...
        "/streaks/milestone/active": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    },
                    {
                        "ApiKey": []
                    }
                ],
                "description": "Returns the streak milestones users can currently reach.\nPaginated with page/page_size; the response is a\ncount/next/previous/results envelope.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "streaks"
                ],
                "summary": "List active streak milestones",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default 10, max 100)",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pagination.PaginatedResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to fetch milestones",
                        "schema": {
                            "$ref": "#/definitions/core.APIError"
                        }
                    }
                }
            }
        },
        "/streaks/milestone/create": {
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    },
                    {
                        "ApiKey": []
                    }
                ],
                "description": "Note: this route only checks IsAuthenticated today, with no admin-style permission gate (see ADR 0006 for the planned fix).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "streaks"
                ],
                "summary": "Create a streak milestone",
                "parameters": [
                    {
                        "description": "Milestone to create",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/repository.CreateStreakMilestoneParams"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/repository.StreakMilestone"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/core.APIError"
                        }
                    },
                    "500": {
                        "description": "Failed to create milestone",
                        "schema": {
                            "$ref": "#/definitions/core.APIError"
                        }
                    }
                }
            }
        },
        "/streaks/milestone/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerToken": []
                    },
                    {
                        "ApiKey": []
                    }
                ],
                "description": "Note: this route only checks IsAuthenticated today, with no admin-style permission gate (see ADR 0006 for the planned fix).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "streaks"
                ],
                "summary": "Delete a streak milestone",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Milestone ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Confirmation message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid id",
                        "schema": {
                            "$ref": "#/definitions/core.APIError"
                        }
                    },
                    "500": {
                        "description": "Failed to delete milestone",
                        "schema": {
                            "$ref": "#/definitions/core.APIError"
                        }
                    }
                }
            }
        },
        "/users/activity/complete": {
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    },
                    {
                        "ApiKey": []
                    }
                ],
                "description": "The request body's account_id must match the caller's own subject.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "streaks"
                ],
                "summary": "Record a completed activity for the authenticated user",
                "parameters": [
                    {
                        "description": "account_id must be the caller's own",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/repository.RecordActivityCompletionParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Confirmation message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/core.APIError"
                        }
                    },
                    "403": {
                        "description": "account_id does not match the caller",
                        "schema": {
                            "$ref": "#/definitions/core.APIError"
                        }
                    },
                    "500": {
                        "description": "Failed to record completion",
                        "schema": {
                            "$ref": "#/definitions/core.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "core.APIError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "the provided input is invalid or malformed"
                }
            }
        },
...
        "pagination.PaginatedResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "next": {
                    "type": "string"
                },
                "previous": {
                    "type": "string"
                },
                "results": {}
            }
        },
...
        "repository.CreateStreakMilestoneParams": {
            "type": "object",
            "properties": {
                "activity_id": {
                    "type": "string"
                },
                "bonus_points": {
                    "type": "integer"
                },
                "days_required": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "repository.RecordActivityCompletionParams": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "string"
                },
                "activity_id": {
                    "type": "string"
                },
                "metadata": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "repository.StreakMilestone": {
            "type": "object",
            "properties": {
                "activity_id": {
                    "type": "string"
                },
                "bonus_points": {
                    "type": "integer"
                },
                "days_required": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKey": {
            "type": "apiKey",
            "name": "X-Api-Key",
            "in": "header"
        },
        "BearerToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}